Move apollo links to their own folder
