+++ src/api/links/session.link.ts
@@ -1,7 +1,7 @@
 import { ApolloClient, fromPromise } from '@apollo/client';
 import { ErrorLink } from '@apollo/client/link/error';
-import { SessionDocument } from '../components/Session/session.generated';
-import { GQLOperations } from './operations.generated';
+import { SessionDocument } from '../../components/Session/session.generated';
+import { GQLOperations } from '../operations.generated';
 
 /**
  * Retry operations on session errors
