+++ src/server/renderServerSideApp.tsx
@@ -16,8 +16,8 @@ import { resetServerContext } from 'react-beautiful-dnd';
 import ReactDOMServer from 'react-dom/server';
 import { FilledContext, HelmetProvider } from 'react-helmet-async';
 import { StaticRouter } from 'react-router-dom/server';
-import { ErrorCache, ErrorCacheLink } from '../api/errorCache.link';
 import { possibleTypes } from '../api/fragmentMatcher.generated';
+import { ErrorCache, ErrorCacheLink } from '../api/links/errorCache.link';
 import { typePolicies } from '../api/typePolicies';
 import { App } from '../App';
 import { Nest } from '../components/Nest';
